requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
jinja2==3.1.4
python-dotenv==1.0.1
opensearch-py==2.7.1
pydantic==2.8.2
httpx==0.27.0
faker==26.0.0
pytest==8.3.2
requests>=2.31.0,<3
pdfminer.six==20231228
pdfplumber==0.11.4
